deep_research_agent/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
